Add files via upload
project involves analyzing a dataset of global average sea level change since 1880 to predict sea level rise through 2050. The data is imported using Pandas, and a scatter plot is created using matplotlib. The linear regression function is used to get the slope and y-intercept of the line of best fit. The line of best fit is plotted over the top of the scatter plot, and a new line is plotted using data from 2000 through the most recent year. The code is written in sea_level_predictor.py, and the project's URL is submitted to freeCodeCamp.